コレクションから検索結果画面

+++ data/src/main/java/com/github/gericass/world_heritage_client/data/AvgleRepositoryImpl.kt
@@ -34,6 +34,7 @@ class AvgleRepositoryImpl(
     }
 
     override suspend fun insertKeyword(keyword: String) {
+        if (keyword.isBlank()) return
         dao.deleteOldest()
         val record = Keyword(keyword, Calendar.getInstance().time)
         dao.insert(record)
